+++ Mobile.BuildTools.sln
@@ -38,6 +38,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mobile.BuildTools.FormsCss"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mobile.BuildTools.Images", "src\Mobile.BuildTools.Images\Mobile.BuildTools.Images.csproj", "{1669A6D6-9B7C-4038-91AE-1851E824C3A1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mobile.BuildTools.Common", "src\Mobile.BuildTools.Common\Mobile.BuildTools.Common.csproj", "{0BF07CD7-3039-44A8-A571-D3B31B40C404}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mobile.BuildTools.Configuration.BuildTasks", "src\Mobile.BuildTools.Configuration.BuildTasks\Mobile.BuildTools.Configuration.BuildTasks.csproj", "{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mobile.BuildTools.AppPlatform", "src\Mobile.BuildTools.AppPlatform\Mobile.BuildTools.AppPlatform.csproj", "{8D1B79C1-024D-4750-87A5-DA62951C79B8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -168,6 +174,42 @@ Global
 		{1669A6D6-9B7C-4038-91AE-1851E824C3A1}.Release|x64.Build.0 = Release|Any CPU
 		{1669A6D6-9B7C-4038-91AE-1851E824C3A1}.Release|x86.ActiveCfg = Release|Any CPU
 		{1669A6D6-9B7C-4038-91AE-1851E824C3A1}.Release|x86.Build.0 = Release|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Debug|x64.Build.0 = Debug|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Debug|x86.Build.0 = Debug|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Release|x64.ActiveCfg = Release|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Release|x64.Build.0 = Release|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Release|x86.ActiveCfg = Release|Any CPU
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404}.Release|x86.Build.0 = Release|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Debug|x64.Build.0 = Debug|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Debug|x86.Build.0 = Debug|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Release|x64.ActiveCfg = Release|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Release|x64.Build.0 = Release|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Release|x86.ActiveCfg = Release|Any CPU
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8}.Release|x86.Build.0 = Release|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Debug|x64.Build.0 = Debug|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Debug|x86.Build.0 = Debug|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Release|x64.ActiveCfg = Release|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Release|x64.Build.0 = Release|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Release|x86.ActiveCfg = Release|Any CPU
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -183,6 +225,9 @@ Global
 		{800F5E43-702F-4309-BA56-411CB20CA595} = {C7456F1C-2B86-4C40-A228-72209A7355A0}
 		{621BA5A5-EA0C-4369-9EE3-99A9794A99E2} = {C7456F1C-2B86-4C40-A228-72209A7355A0}
 		{1669A6D6-9B7C-4038-91AE-1851E824C3A1} = {C7456F1C-2B86-4C40-A228-72209A7355A0}
+		{0BF07CD7-3039-44A8-A571-D3B31B40C404} = {C7456F1C-2B86-4C40-A228-72209A7355A0}
+		{D69D5A9B-CEAC-49FF-B824-CBCBE2C2DCE8} = {C7456F1C-2B86-4C40-A228-72209A7355A0}
+		{8D1B79C1-024D-4750-87A5-DA62951C79B8} = {C7456F1C-2B86-4C40-A228-72209A7355A0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {429C8CA0-DB8C-4CF5-9627-187B9CC47A99}
