+++ src/Mobile.BuildTools.Configuration.BuildTasks/Configuration/MetaData.cs
@@ -1,0 +1,8 @@
+﻿namespace Mobile.BuildTools.Configuration
+{
+    internal static class MetaData
+    {
+        public const string Link = nameof(Link);
+        public const string SourceFile = nameof(SourceFile);
+    }
+}
