+++ src/Mobile.BuildTools.Configuration.BuildTasks/Models/Configuration/ExpectedAppConfigExtensions.cs
