+++ src/Mobile.BuildTools.FormsCss/Mobile.BuildTools.FormsCss.csproj
@@ -16,7 +16,7 @@
     <InternalsVisibleTo Include="Mobile.BuildTools.Tests" />
     <None Include="..\assets\runtimes\**\libsass.*" Pack="true" PackagePath="$(PackFolder)/runtimes/%(RecursiveDir)%(Filename)%(Extension)" />
     <None Include="*.targets;*.props" Pack="true" PackFolder="$(PackFolder)" />
-    <Compile Include="..\Mobile.BuildTools\Logging\BuildHostLoggingHelper.cs" LinkBase="Logging" />
+    <Compile Include="..\Mobile.BuildTools.Common\Logging\BuildHostLoggingHelper.cs" LinkBase="Logging" />
     <None Update="buildTransitive\*.targets;buildTransitive\*.props" Pack="true" PackFolder="buildTransitive" />
   </ItemGroup>
 
@@ -24,10 +24,20 @@
     <PackageReference Include="Microsoft.Build" Pack="false" />
     <PackageReference Include="Microsoft.Build.Tasks.Core" Pack="false" />
     <PackageReference Include="NuGetizer" PrivateAssets="all" />
+    <PackageReference Include="System.Text.Json" PrivateAssets="All" />
+    <PackageReference Include="System.ComponentModel.Annotations" PrivateAssets="all" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Mobile.BuildTools.Reference\Mobile.BuildTools.Reference.csproj" PackFolder="build" />
   </ItemGroup>
+
+  <Target Name="ExcludeTFM"
+          BeforeTargets="_AddPackageManifest"
+          AfterTargets="InferPackageContents">
+    <PropertyGroup>
+      <IsPackagingProject>true</IsPackagingProject>
+    </PropertyGroup>
+  </Target>
 
 </Project>
