+++ src/Mobile.BuildTools.Images/buildTransitive/Mobile.BuildTools.Images.targets
@@ -1,0 +1,3 @@
+<Project>
+  <Import Project="$(MSBuildThisFileDirectory)..\build\Mobile.BuildTools.Images.targets" />
+</Project>
