+++ src/Mobile.BuildTools/Mobile.BuildTools.csproj
@@ -15,7 +15,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\Mobile.BuildTools.Configuration\TransformationHelper.cs" Link="Utils\TransformationHelper.cs" />
     <InternalsVisibleTo Include="Mobile.BuildTools.Tests" />
   </ItemGroup>
 
@@ -27,17 +26,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="HarfBuzzSharp" PrivateAssets="all" />
+    <!--<PackageReference Include="HarfBuzzSharp" PrivateAssets="all" />-->
     <PackageReference Include="Microsoft.Build" Pack="false" />
     <PackageReference Include="Microsoft.Build.Tasks.Core" Pack="false" />
     <PackageReference Include="NuGetizer" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.Web.Xdt" PrivateAssets="all" />
-    <PackageReference Include="ShimSkiaSharp" PrivateAssets="all" />
-    <PackageReference Include="Svg.Skia" PrivateAssets="all" />
-    <PackageReference Include="System.Buffers" PrivateAssets="all" />
+    <!--<PackageReference Include="System.Buffers" PrivateAssets="all" />
     <PackageReference Include="System.IO.UnmanagedMemoryStream" PrivateAssets="all" />
     <PackageReference Include="System.Numerics.Vectors" PrivateAssets="all" />
-    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" PrivateAssets="all" />
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" PrivateAssets="all" />-->
     <PackageReference Include="System.Text.Json" PrivateAssets="all" />
   </ItemGroup>
 
@@ -49,7 +45,6 @@
     <ProjectReference Include="..\Mobile.BuildTools.Images\Mobile.BuildTools.Images.csproj"
                       ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\Mobile.BuildTools.Reference\Mobile.BuildTools.Reference.csproj" PackFolder="build" />
-    <ProjectReference Include="..\Mobile.BuildTools.XamarinSdk\Mobile.BuildTools.XamarinSdk.csproj" PackFolder="build" />
   </ItemGroup>
 
   <Target Name="ExcludeTFM"
